Drop `support`, reorg a little, tests
