--- snippets/tests/it/language/c99_tc3/testdata/hello_world.cpp
@@ -1,7 +1,0 @@
-
-#include <iostream>
-
-int main() {
-    std::cout << "Hello World!";
-    return 0;
-}
