+++ snippets/tests/it/language/testdata/c99_tc3/hello_world.c
