+++ snippets/tests/it/language/testdata/c99_tc3/hello_world_error.c
@@ -1,0 +1,7 @@
+
+#include <stdio.h>
+
+int main() {
+  printf("hello world
+  return 0;
+}
